Removed old script
Removed old script of displaying system time on html screen. Now the website is working properly by displaying time on the screen. 
But now changes need to done are,
Remove the script that is directly written on the html page inside Script tag.
And place it inside script.js file